Leaks: introduce a 'Disposable' interface and use it to clean up. (#1115)
This commit fixes several classes to be more careful with their
lifetimes and event listeners in order to minimize leaks.

Because the audio-video controller can be stopped and restarted, as can
other classes, a `Disposable` interface is introduced in this commit.
When you are completely finished with meeting objects, call `dispose()`
and discard those objects.

Additionally, this commit causes the `ended` event to be dispatched on
`MediaStreamTrack`s that are closed by the SDK itself.